main140820231335 fuctionality of font size changer implimented,small bug fix
now as per the font size giving in the font field our entire content will change its font resepctive to it. Its because this font is taken as an reference px for the children. Update the code to fix an bug, that if i add any other tag theres comes an error due to if there is already an header tag (error in accessing ankerId)